+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import CalendarCalculator from './App';
+import App from './App';
 
-ReactDOM.render(<CalendarCalculator />, document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
 // ReactDOM.render(<h1>test render</h1>, document.getElementById('root'));
